requirements.txt
streamlit
asyncio
pandas
tiktoken
langchain
langchain-community
langchain_openai
openai
openpyxl
xlsxwriter
beautifulsoup4
markdown
aiohttp
python-dotenv
watchdog
ollama
setuptools
requests
google_auth_oauthlib
google-auth-httplib2 
google-api-python-client
google-generativeai
langchain-google-genai
scrapeless